fix: cleanup handleClickableElement
