+++ src/modules/colorReflection.ts
@@ -1,9 +1,8 @@
 export const colorReflection = (
-  reflexOne: HTMLElement | null,
-  reflexTwo: HTMLElement | null,
+  reflexOne: HTMLElement,
+  reflexTwo: HTMLElement,
   color: string,
 ): void => {
-  if (reflexOne === null || reflexTwo === null) return;
   reflexOne.style.backgroundColor = color;
   reflexTwo.style.backgroundColor = color;
 };
