+++ src/modules/index.ts
@@ -1,3 +1,4 @@
 export * from './generateBackgroundColor';
 export * from './colorSVG';
 export * from './colorReflection';
+export * from './handleClickableElement';
